Changes to Player Model
Ran into an issue where I missed a step from the start of the model, I tried adding it in and then making small changes to get the model looking good but was unable to. I am starting over fram scratch and will make sure to follow all steps this time.